Group all bad_word into profanity
