+++ tests/fixtures/texts/profanities.js
@@ -1,4 +1,42 @@
-const getTextBadWord = `
+const retrieveFilter = `
+{
+  "data": {
+    "id": "5aaf6fec6e1157cad6d66b50",
+    "project_id": 0,
+    "filter_set": [
+      "foo",
+      "bar",
+      "baz"
+    ],
+    "use_default": true
+  },
+  "meta": {
+    "code": 200,
+    "message": "success"
+  }
+}
+`;
+
+const createFilter = `
+{
+  "data": {
+    "id": "5aaf72006e1157cad6d66b51",
+    "project_id": 114,
+    "filter_set": [
+      "foo",
+      "bar",
+      "baz"
+    ],
+    "use_default": false
+  },
+  "meta": {
+    "code": 200,
+    "message": "success"
+  }
+}
+`;
+
+const moderationResult = `
 {
   "data": [
     {
@@ -18,7 +56,7 @@ const getTextBadWord = `
 }
 `;
 
-const createTextBadWord = `
+const sanitize = `
 {
   "data": {
     "id": "5abb1cea6e1157d322ef7a2a",
@@ -37,6 +75,8 @@ const createTextBadWord = `
 `;
 
 module.exports = {
-  getTextBadWord,
-  createTextBadWord,
+  retrieveFilter,
+  createFilter,
+  moderationResult,
+  sanitize,
 };
